backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.1.2
openai>=1.3.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-jose==3.3.0
gunicorn==21.2.0
SQLAlchemy==2.0.27 